[clean]クイズ作成/更新ボタンのメソッドを別ファイルへ移動 #7
クイズ作成と更新ボタンの記述がcreateQuiz.tsxに記述されており、
記述量が増え可読性が著しく下がっていたため、作成/更新ボタンの処理を別ファイルに移した。
同じファイルに作成/更新ボタンの記述が存在する理由は、
handleUpdateQuizDataメソッド内でhandleCreateQuizSubmitメソッドを使用する場面があり、別のファイルにしてしまうと使用できなくなってしまうため。
