+++ frontend/app/src/components/pages/CreateQuiz.tsx
@@ -2,30 +2,28 @@ import React, { useContext, useState, createContext, useEffect } from "react";
 import { useHistory, useLocation, useParams } from "react-router-dom";
 
 import { makeStyles, Theme } from "@material-ui/core";
-import Button from "@material-ui/core/Button"
 
 import AboutQuiz from "./AboutQuiz";
 import { AuthContext } from "App";
-import { AboutQuizData, QuizFirtsOrLastData } from "interfaces";
-import { createQuiz, getQuiz } from "lib/api/quizzes";
+
+import { getQuiz } from "lib/api/quizzes";
 import Terminal from "./Terminal";
 import QuizBranchArea from "./QuizBranchArea";
-import { createQuizFirstOrLast, getQuizFirstOrLast } from "lib/api/quiz_first_or_lasts";
-import { createQuizBranch, deleteQuizBranch, getQuizBranch, updateQuizBranch } from "lib/api/quiz_branches";
-import { createQuizWorktreeFile, deleteQuizWorktreeFile, updateQuizWorktreeFile } from "lib/api/quiz_worktree_files";
-import { createQuizCommitMessage, deleteQuizCommitMessage, getQuizCommitMessage } from "lib/api/quiz_commit_messages";
-import { createQuizRepositoryFile, deleteQuizRepositoryFile } from "lib/api/quiz_repository_files";
-import { createQuizIndexFile, deleteQuizIndexFile, updateQuizIndexFile } from "lib/api/quiz_index_files";
-import { CreateQuizHistoryOfCommittedFile, deleteQuizHistoryOfCommittedFile } from "lib/api/quiz_history_of_committed_files";
-import { createQuizRemoteBranch, deleteQuizRemoteBranch, getQuizRemoteBranch } from "lib/api/quiz_remote_branches";
-import { createQuizRemoteCommitMessage, deleteQuizRemoteCommitMessage, getQuizRemoteCommitMessage } from "lib/api/quiz_remote_commit_messages";
-import { createQuizRemoteRepositoryFile, deleteQuizRemoteRepositoryFile } from "lib/api/quiz_remote_repository_file";
+import { getQuizFirstOrLast } from "lib/api/quiz_first_or_lasts";
+import { getQuizBranch } from "lib/api/quiz_branches";
+
+import { getQuizCommitMessage } from "lib/api/quiz_commit_messages";
+import { getQuizRemoteBranch } from "lib/api/quiz_remote_branches";
+import { getQuizRemoteCommitMessage } from "lib/api/quiz_remote_commit_messages";
+import CreateOrUpdateQuizButton from "./CreateQuizButton";
 
 export const QuizContext = createContext({} as {
   quizTitle: string
   setQuizTitle: React.Dispatch<React.SetStateAction<string>>
   quizIntroduction: string
   setQuizIntroduction: React.Dispatch<React.SetStateAction<string>>
+  quizFirstOrLastId: string
+  setQuizFirstOrLastId: React.Dispatch<React.SetStateAction<string>>
   text: string
   setText: React.Dispatch<React.SetStateAction<string>>
   addText: string
@@ -346,19 +344,7 @@ export const QuizContext = createContext({} as {
   }[]>>
   })
 
-const useStyles = makeStyles((theme: Theme) => ({
-  submitBtn: {
-    marginTop: theme.spacing(2),
-    flexGrow: 1,
-    textTransform: "none",
-    backgroundColor: "#f5f5f5"
-  }
-}))
-
 const CreateQuiz: React.FC = () => {
-  const { currentUser } = useContext(AuthContext)
-  const history = useHistory()
-  const classes = useStyles()
   const location = useLocation()
   const { id } = useParams<{ id: string }>()
 
@@ -505,223 +491,12 @@ const CreateQuiz: React.FC = () => {
     addText:""
   }]);
 
-  const quizType = "user"
-
-  const aboutQuizData: AboutQuizData | undefined  =
-  location.pathname === "/quiz"
-  ? {
-      quizTitle: quizTitle,
-      quizIntroduction: quizIntroduction,
-      quizType: quizType,
-      userId: currentUser?.id
-    }
-  : undefined
-
-  const handleCreateQuizSubmit = (argBranches :any, argsWorktreeFiles :any, argIndexFiles :any, argCommitMessages :any, argFileHistoryForCansellCommits :any, argRepositoryFiles :any, argRemoteBranches :any, argRemoteCommitMessages :any, argRemoteRepositoryFiles :any) => async (e: React.MouseEvent<HTMLButtonElement>) => {
-    e.preventDefault()
-    const quizFirtsOrLastStatus = "last"
-
-    try {
-      const aboutQuizRes = await createQuiz(aboutQuizData)
-
-      const quizFirtsOrLastData: QuizFirtsOrLastData | undefined =
-      location.pathname === "/quiz"
-      ? {
-          quizFirstOrLastStatus: quizFirtsOrLastStatus,
-          quizId: aboutQuizRes?.data.data.id
-        }
-      : undefined
-
-      const quizFirtsOrLastRes = await createQuizFirstOrLast(quizFirtsOrLastData)
-
-      const quizBranchData = argBranches.map((branch :any) => ({
-        quizBranchName: branch.branchName,
-        quizFirstOrLastId: location.pathname === "/quiz" ? quizFirtsOrLastRes?.data.data.id : quizFirstOrLastId
-      }))
-
-      const quizBranchRes = await createQuizBranch(quizBranchData)
-
-      const quizWorktreeFileData = quizBranchRes !== undefined
-      ? quizBranchRes.data.data.map((branch :any) =>(
-          argsWorktreeFiles
-          .filter((worktreeFile :any) => worktreeFile.parentBranch === branch.quizBranchName)
-          .map((filteredWorktreeFile :any) =>({
-            quizWorktreeFileName: filteredWorktreeFile.fileName,
-            quizWorktreeFileTextStatus: filteredWorktreeFile.textStatus,
-            quizBranchId: branch.id
-          }))
-        ))
-      : []
-
-      const quizWorktreeFileRes = await createQuizWorktreeFile(quizWorktreeFileData.flat())
-
-      const quizIndexFileData = quizBranchRes !== undefined
-      ? quizBranchRes.data.data.map((branch :any) =>(
-          argIndexFiles
-          .filter((indexFile :any) => indexFile.parentBranch === branch.quizBranchName)
-          .map((filteredIndexFile :any) =>({
-            quizIndexFileName: filteredIndexFile.fileName,
-            quizIndexFileTextStatus: filteredIndexFile.textStatus,
-            quizBranchId: branch.id
-          }))
-        ))
-      : []
-
-      const quizIndexFileRes = await createQuizIndexFile(quizIndexFileData.flat())
-
-      const QuizCommitMessageData = quizBranchRes !== undefined
-      ? quizBranchRes.data.data.map((branch :any) =>(
-          argCommitMessages
-          .filter((commitMessage :any) => commitMessage.parentBranch === branch.quizBranchName)
-          .map((commitMessage :any) =>({
-            quizCommitMessage: commitMessage.message,
-            quizBranchId: branch.id
-          }))
-        ))
-      : []
-
-      const QuizCommitMessageRes = await createQuizCommitMessage(QuizCommitMessageData.flat())
-
-      const quizHistoryOfCommittedFileData = QuizCommitMessageRes !== undefined
-      ? QuizCommitMessageRes.data.data.map((commitMessage :any) =>(
-        argFileHistoryForCansellCommits
-        .filter((historyFile :any) =>
-          historyFile.parentCommitMessage === commitMessage.quizCommitMessage
-          && historyFile.fileName !== "")
-        .map((filteredHistoryFile :any) =>({
-          quizHistoryOfCommittedFileName: filteredHistoryFile.fileName,
-          quizHistoryOfCommittedFileStatus: filteredHistoryFile.fileStatus,
-          quizHistoryOfCommittedFileTextStatus: filteredHistoryFile.textStatus,
-          quizHistoryOfCommittedFilePastTextStatus: filteredHistoryFile.pastTextStatus,
-          quizHistoryOfCommittedFileParentPastCommitMessage: filteredHistoryFile.parentPastCommitMessage,
-          quizCommitMessageId: commitMessage.id,
-          quizBranchId: commitMessage.quizBranchId,
-        }))
-      ))
-      : []
-
-      const flatedQuizHistoryOfCommittedFileData = quizHistoryOfCommittedFileData.flat()
-
-      const removeDuplicateQuizHistoryOfCommittedFileData = flatedQuizHistoryOfCommittedFileData.filter((e :any, index :any, self :any) => {
-        return self.findIndex((el :any) =>
-          el.quizHistoryOfCommittedFileName === e.quizHistoryOfCommittedFileName
-          && el.quizHistoryOfCommittedFileTextStatus === e.quizHistoryOfCommittedFileTextStatus
-          && el.quizHistoryOfCommittedFilePastTextStatus === e.quizHistoryOfCommittedFilePastTextStatus
-          && el.quizHistoryOfCommittedFileParentPastCommitMessage === e.quizHistoryOfCommittedFileParentPastCommitMessage
-          && el.quizCommitMessageId === e.quizCommitMessageId
-          && el.quizBranchId === e.quizBranchId
-        ) === index;
-      });
-
-      const quizHistoryOfCommittedFileRes = await CreateQuizHistoryOfCommittedFile(removeDuplicateQuizHistoryOfCommittedFileData)
-
-      const quizRepositoryFileData = QuizCommitMessageRes !== undefined
-      ? QuizCommitMessageRes.data.data.map((commitMessage :any) =>(
-        argRepositoryFiles
-        .filter((repositoryFile :any) =>
-          repositoryFile.parentCommitMessage === commitMessage.quizCommitMessage
-          && repositoryFile.fileName !== "")
-        .map((filteredRepositoryFile :any) =>({
-          quizRepositoryFileName: filteredRepositoryFile.fileName,
-          quizRepositoryFileTextStatus: filteredRepositoryFile.textStatus,
-          quizCommitMessageId: commitMessage.id,
-          quizBranchId: commitMessage.quizBranchId
-        }))
-      ))
-      : []
-
-      const flatedQuizRepositoryFileData = quizRepositoryFileData.flat()
-
-      const removeDuplicateQuizRepositoryFileData = flatedQuizRepositoryFileData.filter((e :any, index :any, self :any) => {
-        return self.findIndex((el :any) =>
-          el.quizRepositoryFileName === e.quizRepositoryFileName
-          && el.quizRepositoryFileTextStatus === e.quizRepositoryFileTextStatus
-          && el.quizCommitMessageId === e.quizCommitMessageId
-          && el.quizBranchId === e.quizBranchId
-        ) === index;
-      });
-
-      const quizRepositoryFileRes = await createQuizRepositoryFile(removeDuplicateQuizRepositoryFileData)
-
-      const quizRemoteBranchData =
-        argRemoteBranches
-        .filter((branch :any) => branch.remoteBranchName)
-        .map((branch :any) => ({
-          quizRemoteBranchName: branch.remoteBranchName,
-          quizFirstOrLastId: quizFirtsOrLastRes === undefined ? quizFirstOrLastId : quizFirtsOrLastRes.data.data.id
-        }
-      ))
-
-      const quizRemoteBranchRes = await createQuizRemoteBranch(quizRemoteBranchData)
-
-      const QuizRemoteCommitMessageData = quizRemoteBranchRes !== undefined
-      ? quizRemoteBranchRes.data.data.map((branch :any) =>(
-          argRemoteCommitMessages
-          .filter((commitMessage :any) => commitMessage.parentRemoteBranch === branch.quizRemoteBranchName)
-          .map((commitMessage :any) =>({
-            quizRemoteCommitMessage: commitMessage.remoteMessage,
-            quizRemoteBranchId: branch.id
-          }))
-        ))
-      : []
-
-      const QuizRemoteCommitMessageRes = await createQuizRemoteCommitMessage(QuizRemoteCommitMessageData.flat())
-
-      const quizRemoteRepositoryFileData = QuizRemoteCommitMessageRes !== undefined
-      ? QuizRemoteCommitMessageRes.data.data.map((commitMessage :any) =>(
-        argRemoteRepositoryFiles
-        .filter((repositoryFile :any) =>
-          repositoryFile.parentRemoteCommitMessage === commitMessage.quizRemoteCommitMessage
-          && repositoryFile.fileName !== "")
-        .map((filteredRepositoryFile :any) =>({
-          quizRemoteRepositoryFileName: filteredRepositoryFile.fileName,
-          quizRemoteRepositoryFileTextStatus: filteredRepositoryFile.textStatus,
-          quizRemoteCommitMessageId: commitMessage.id,
-          quizRemoteBranchId: commitMessage.quizRemoteBranchId
-        }))
-      ))
-      : []
-
-      const flatedQuizRemoteRepositoryFileData = quizRemoteRepositoryFileData.flat()
-
-      const removeDuplicateQuizRemoteRepositoryFileData = flatedQuizRemoteRepositoryFileData.filter((e :any, index :any, self :any) => {
-        return self.findIndex((el :any) =>
-          el.quizRemoteRepositoryFileName === e.quizRemoteRepositoryFileName
-          && el.quizRemoteRepositoryFileTextStatus === e.quizRemoteRepositoryFileTextStatus
-          && el.quizRemoteCommitMessageId === e.quizRemoteCommitMessageId
-          && el.quizRemoteBranchId === e.quizRemoteBranchId
-        ) === index;
-      });
-
-      const quizRemoteRepositoryFileRes = await createQuizRemoteRepositoryFile(removeDuplicateQuizRemoteRepositoryFileData)
-
-      console.log(aboutQuizRes)
-      console.log(quizFirtsOrLastRes)
-      console.log(quizBranchRes)
-      console.log(quizWorktreeFileRes)
-      console.log(quizIndexFileRes)
-      console.log(QuizCommitMessageRes)
-      console.log(quizHistoryOfCommittedFileRes)
-      console.log(quizRepositoryFileRes)
-      console.log(quizRemoteBranchRes)
-      console.log(QuizRemoteCommitMessageRes)
-      console.log(quizRemoteRepositoryFileRes)
-      console.log("create quiz success!!")
-
-      history.push("/")
-
-    } catch (err) {
-      console.log(err)
-    }
-  }
-
   const addBranches = (array :any, branchName :any, branchId :any) => {
     array((branches :any) => [...branches,{
       branchName: branchName,
       branchId: branchId
     }])
   }
-
   const addWorktreeFiles = (array :any, fileName :any, parentBranch :any, textStatus :any, parentBranchId :any,worktreeFileId :any) => {
     array((worktreeFile :any) => [...worktreeFile,{
       fileName: fileName,
@@ -731,7 +506,6 @@ const CreateQuiz: React.FC = () => {
       worktreeFileId: worktreeFileId
     }])
   }
-
   const addIndexFile = (array :any, fileName :any, parentBranch :any, textStatus :any, parentBranchId :any, worktreeFileId :any) => {
     array((indexFile :any) => [...indexFile,{
       fileName: fileName,
@@ -741,7 +515,6 @@ const CreateQuiz: React.FC = () => {
       indexFileId: worktreeFileId
     }])
   }
-
   const addCommitMessages = (array :any, message :any, parentBranch :any, parentBranchId :any, commitMessageId :any) => {
     array((commitMessage :any) => [...commitMessage,{
       message: message,
@@ -750,7 +523,6 @@ const CreateQuiz: React.FC = () => {
       commitMessageId: commitMessageId
     }])
   }
-
   const addRepositoryFiles = (array :any, fileName :any, textStatus :any, parentBranch :any, parentCommitMessage :any, parentBranchId :any, parentCommitMessageId :any, repositoryFileId: any) => {
     array((repositoryFile :any) => [...repositoryFile,{
       fileName: fileName,
@@ -762,7 +534,6 @@ const CreateQuiz: React.FC = () => {
       repositoryFileId: repositoryFileId
     }])
   }
-
   const addFileHistoryForCansellCommits = (array :any, fileName :any, fileStatus :any, textStatus :any, pastTextStatus :any, parentBranch :any, parentCommitMessage :any, parentPastCommitMessage :any, parentBranchId :any, parentCommitMessageId :any, historyFileId: any) => {
     array((historyFiles :any) => [...historyFiles,{
       fileName: fileName,
@@ -777,14 +548,12 @@ const CreateQuiz: React.FC = () => {
       historyFileId: historyFileId
     }])
   }
-
   const addRemoteBranches = (array :any, remoteBranchName :any, remoteBranchId :any) => {
     array((branches :any) => [...branches,{
       remoteBranchName: remoteBranchName,
       remoteBranchId: remoteBranchId
     }])
   }
-
   const addRemoteCommitMessages = (array :any, remoteMessage :any, parentRemoteBranch :any, parentRemoteBranchId :any, remoteCommitMessageId :any) => {
     array((commitMessage :any) => [...commitMessage,{
       remoteMessage: remoteMessage,
@@ -793,7 +562,6 @@ const CreateQuiz: React.FC = () => {
       remoteCommitMessageId: remoteCommitMessageId
     }])
   }
-
   const addRemoteRepositoryFiles = (array :any, fileName :any, textStatus :any, parentRemoteBranch :any, parentRemoteCommitMessage :any, parentRemoteBranchId :any, parentRemoteCommitMessageId :any, remoteRepositoryFileId: any) => {
     array((repositoryFile :any) => [...repositoryFile,{
       fileName: fileName,
@@ -958,265 +726,6 @@ const CreateQuiz: React.FC = () => {
     }
   }
 
-  //ブランチidと自身のidを持っていない配列
-  const createBranches = branches.filter(branch => branch.branchName && !branch.branchId)
-  const createWorktreeFiles = worktreeFiles.filter(worktreeFile => worktreeFile.fileName && !worktreeFile.worktreeFileId && !worktreeFile.parentBranchId)
-  const createIndexFiles = indexFiles.filter(indexFile => indexFile.fileName && !indexFile.indexFileId && !indexFile.parentBranchId)
-  const createRepositoryFiles = repositoryFiles.filter(repositoryFile => repositoryFile.fileName && !repositoryFile.repositoryFileId && !repositoryFile.parentBranchId && !repositoryFile.parentCommitMessageId)
-  const createCommitMessages = commitMessages.filter(commitMessage => commitMessage.message && !commitMessage.commitMessageId && !commitMessage.parentBranchId)
-  const createFileHistoryForCansellCommits = fileHistoryForCansellCommits.filter(historyFile => historyFile.fileName && !historyFile.historyFileId && !historyFile.parentBranchId && !historyFile.parentCommitMessageId)
-  const createRemoteBranches = remoteBranches.filter(remoteBranch => remoteBranch.remoteBranchName && !remoteBranch.remoteBranchId)
-  const createRemoteCommitMessages = remoteCommitMessages.filter(remoteCommitMessage => remoteCommitMessage.remoteMessage && !remoteCommitMessage.remoteCommitMessageId && !remoteCommitMessage.parentRemoteBranchId)
-  const createRemoteRepositoryFiles = remoteRepositoryFiles.filter(remoteRepositoryFile => remoteRepositoryFile.fileName && !remoteRepositoryFile.remoteRepositoryFileId && !remoteRepositoryFile.parentRemoteBranchId && !remoteRepositoryFile.parentRemoteCommitMessageId)
-
-  //ブランチidは持っていて自身のidを持っていない配列
-  const createNewBranchWorktreeFiles =
-    worktreeFiles
-    .filter(worktreeFile => !worktreeFile.worktreeFileId && worktreeFile.parentBranchId)
-    .map(filteredWorktreeFile =>({
-      quizWorktreeFileName: filteredWorktreeFile.fileName,
-      quizWorktreeFileTextStatus: filteredWorktreeFile.textStatus,
-      quizBranchId: filteredWorktreeFile.parentBranchId
-    }))
-  const createNewBranchIndexFiles =
-    indexFiles
-    .filter(indexFile => !indexFile.indexFileId && indexFile.parentBranchId)
-    .map(filteredIndexFile =>({
-      quizIndexFileName: filteredIndexFile.fileName,
-      quizIndexFileTextStatus: filteredIndexFile.textStatus,
-      quizBranchId: filteredIndexFile.parentBranchId
-    }))
-  const createNewBranchCommitMessages =
-    commitMessages
-    .filter(commitMessage => !commitMessage.commitMessageId && commitMessage.parentBranchId)
-    .map((filtedcommitMessage :any) =>({
-      quizCommitMessage: filtedcommitMessage.message,
-      quizBranchId: filtedcommitMessage.parentBranchId
-    }))
-  const createNewBranchRemoteCommitMessages =
-    remoteCommitMessages
-    .filter(remoteCommitMessage => !remoteCommitMessage.remoteCommitMessageId && remoteCommitMessage.parentRemoteBranchId)
-    .map((filtedcommitMessage :any) =>({
-      quizRemoteCommitMessage: filtedcommitMessage.remoteMessage,
-      quizRemoteBranchId: filtedcommitMessage.parentRemoteBranchId
-    }))
-
-  const updateBranches =
-    branches.filter(branch => initialBranches.some(initBranch =>
-      branch.branchName !== initBranch.branchName
-      && branch.branchId === initBranch.branchId
-      && branch.branchId
-      && branch.branchName
-    ))
-  const updateWorktreeFiles =
-    worktreeFiles.filter(worktreeFile => initialWorktreeFiles.some(initWorkfile =>
-      initWorkfile.fileName === worktreeFile.fileName
-      && initWorkfile.parentBranch === worktreeFile.parentBranch
-      && initWorkfile.textStatus !== worktreeFile.textStatus
-      && initWorkfile.worktreeFileId === worktreeFile.worktreeFileId
-    ))
-  const updateIndexFiles =
-    indexFiles.filter(indexFile => initialIndexFiles.some(initialIndexFile =>
-      indexFile.fileName === initialIndexFile.fileName
-      && indexFile.parentBranch === initialIndexFile.parentBranch
-      && indexFile.textStatus !== initialIndexFile.textStatus
-      && indexFile.indexFileId === initialIndexFile.indexFileId
-    ))
-
-  const deleteBranches =
-    initialBranches.filter(initialBranch => !branches.some(branch =>
-      initialBranch.branchId === branch.branchId
-    ))
-  const deleteCommitMessages =
-    initialCommitMessages.filter(initCommit =>
-      !commitMessages.some(commitMessage => initCommit.commitMessageId === commitMessage.commitMessageId)
-      && !deleteBranches.some(deleteBranch => initCommit.parentBranch === deleteBranch.branchName)
-    )
-  const deleteWorktreeFiles =
-    initialWorktreeFiles.filter(initWorkfile =>
-      !worktreeFiles.some(worktreeFile => initWorkfile.worktreeFileId === worktreeFile.worktreeFileId)
-      && !deleteBranches.some(deleteBranch => initWorkfile.parentBranch === deleteBranch.branchName)
-    )
-  const deleteIndexFiles =
-    initialIndexFiles.filter(initIndexFile => !indexFiles.some(indexFile =>
-      initIndexFile.indexFileId === indexFile.indexFileId
-      && !deleteBranches.some(deleteBranch => initIndexFile.parentBranch === deleteBranch.branchName)
-    ))
-  const deleteRepositoryFiles =
-    initialRepositoryFiles.filter(initialRepositoryFile => !repositoryFiles.some(repositoryFile =>
-      initialRepositoryFile.repositoryFileId === repositoryFile.repositoryFileId
-      && !deleteBranches.some(deleteBranch => initialRepositoryFile.parentBranch === deleteBranch.branchName)
-      && !deleteCommitMessages.some(deleteCommitMessage =>  initialRepositoryFile.parentCommitMessage === deleteCommitMessage.message)
-    ))
-  const deleteFileHistoryForCansellCommits =
-    initialFileHistoryForCansellCommits.filter(initialFileHistoryForCansellCommit => !fileHistoryForCansellCommits.some(fileHistoryForCansellCommit =>
-      initialFileHistoryForCansellCommit.historyFileId === fileHistoryForCansellCommit.historyFileId
-      && !deleteBranches.some(deleteBranch => initialFileHistoryForCansellCommit.parentBranch === deleteBranch.branchName)
-      && !deleteCommitMessages.some(deleteCommitMessage =>  initialFileHistoryForCansellCommit.parentCommitMessage === deleteCommitMessage.message)
-    ))
-  const deleteRemoteBranches =
-    initialRemoteBranches.filter(initialRemoteBranch => !remoteBranches.some(remoteBranch =>
-      initialRemoteBranch.remoteBranchId === remoteBranch.remoteBranchId
-    ))
-  const deleteRemoteCommitMessages =
-    initialRemoteCommitMessages.filter(initialRemoteCommitMessage => !remoteCommitMessages.some(remoteCommitMessage =>
-      initialRemoteCommitMessage.remoteCommitMessageId === remoteCommitMessage.remoteCommitMessageId
-      && !deleteRemoteBranches.some(deleteRemoteBranches => initialRemoteCommitMessage.parentRemoteBranch === deleteRemoteBranches.remoteBranchName)
-    ))
-  const deleteRemoteRepositoryFiles =
-    initialRemoteRepositoryFiles.filter(initialRemoteRepositoryFile => !remoteRepositoryFiles.some(remoteRepositoryFile =>
-      remoteRepositoryFile.remoteRepositoryFileId === initialRemoteRepositoryFile.remoteRepositoryFileId
-      && !deleteRemoteBranches.some(deleteRemoteBranches => initialRemoteRepositoryFile.parentRemoteBranch === deleteRemoteBranches.remoteBranchName)
-      && !deleteCommitMessages.some(deleteCommitMessage =>  initialRemoteRepositoryFile.parentRemoteCommitMessage === deleteCommitMessage.message)
-    ))
-
-  const handleUpdateQuizData = async (e: React.MouseEvent<HTMLButtonElement>) => {
-    try {
-
-      //削除
-      console.log("deleteBranches", deleteBranches)
-      console.log("deleteCommitMessages", deleteCommitMessages)
-      console.log("deleteWorktreeFiles", deleteWorktreeFiles)
-      console.log("deleteIndexFiles", deleteIndexFiles)
-      console.log("deleteRepositoryFiles", deleteRepositoryFiles)
-      console.log("deleteFileHistoryForCansellCommit", deleteFileHistoryForCansellCommits)
-      console.log("deleteRemoteBranches", deleteRemoteBranches)
-
-      //新規作成用の関数(branchも新規作成の場合)
-      await handleCreateQuizSubmit(
-        createBranches,
-        createWorktreeFiles,
-        createIndexFiles,
-        createCommitMessages,
-        createFileHistoryForCansellCommits,
-        createRepositoryFiles,
-        createRemoteBranches,
-        createRemoteCommitMessages,
-        createRemoteRepositoryFiles
-      )(e)
-      // 既存ブランチ内の新規作成
-      const quizWorktreeFileRes = await createQuizWorktreeFile(createNewBranchWorktreeFiles)
-      const quizIndexFileRes = await createQuizIndexFile(createNewBranchIndexFiles)
-      const QuizCommitMessageRes = await createQuizCommitMessage(createNewBranchCommitMessages)
-
-      const createNewBranchFileHistoryForCansellCommits =
-        QuizCommitMessageRes !== undefined
-        ? QuizCommitMessageRes.data.data.map((commitMessage :any) =>(
-            fileHistoryForCansellCommits
-            .filter((historyFile :any) =>
-              historyFile.parentCommitMessage === commitMessage.quizCommitMessage
-              && historyFile.fileName !== ""
-              && historyFile.parentBranchId
-              && !historyFile.historyFileId
-              && !historyFile.parentCommitMessageId)
-            .map((filteredHistoryFile :any) =>({
-              quizHistoryOfCommittedFileName: filteredHistoryFile.fileName,
-              quizHistoryOfCommittedFileStatus: filteredHistoryFile.fileStatus,
-              quizHistoryOfCommittedFileTextStatus: filteredHistoryFile.textStatus,
-              quizHistoryOfCommittedFilePastTextStatus: filteredHistoryFile.pastTextStatus,
-              quizHistoryOfCommittedFileParentPastCommitMessage: filteredHistoryFile.parentPastCommitMessage,
-              quizCommitMessageId: commitMessage.id,
-              quizBranchId: filteredHistoryFile.parentBranchId,
-            }))
-          ))
-        : []
-      const quizHistoryOfCommittedFileRes = await CreateQuizHistoryOfCommittedFile(createNewBranchFileHistoryForCansellCommits.flat())
-
-      const createNewBranchRepositoryFiles =
-        QuizCommitMessageRes !== undefined
-        ? QuizCommitMessageRes.data.data.map((commitMessage :any) =>(
-            repositoryFiles
-            .filter((repositoryFile :any) =>
-              repositoryFile.parentCommitMessage === commitMessage.quizCommitMessage
-              && repositoryFile.fileName !== ""
-              && repositoryFile.parentBranchId
-              && !repositoryFile.repositoryFileId
-              && !repositoryFile.parentCommitMessageId)
-            .map((filteredRepositoryFile :any) =>({
-              quizRepositoryFileName: filteredRepositoryFile.fileName,
-              quizRepositoryFileTextStatus: filteredRepositoryFile.textStatus,
-              quizCommitMessageId: commitMessage.id,
-              quizBranchId: filteredRepositoryFile.parentBranchId
-            }))
-          ))
-        : []
-      const quizRepositoryFileRes = await createQuizRepositoryFile(createNewBranchRepositoryFiles.flat())
-
-      const QuizRemoteCommitMessageRes = await createQuizRemoteCommitMessage(createNewBranchRemoteCommitMessages)
-
-      const createNewBranchRemoteRepositoryFiles =
-      QuizRemoteCommitMessageRes !== undefined
-      ? QuizRemoteCommitMessageRes.data.data.map((commitMessage :any) =>(
-          remoteRepositoryFiles
-          .filter((repositoryFile :any) =>
-            repositoryFile.parentRemoteCommitMessage === commitMessage.quizRemoteCommitMessage
-            && repositoryFile.fileName !== ""
-            && repositoryFile.parentRemoteBranchId
-            && !repositoryFile.remoteRepositoryFileId
-            && !repositoryFile.parentRemoteCommitMessageId)
-          .map((filteredRepositoryFile :any) =>({
-            quizRemoteRepositoryFileName: filteredRepositoryFile.fileName,
-            quizRemoteRepositoryFileTextStatus: filteredRepositoryFile.textStatus,
-            quizRemoteCommitMessageId: commitMessage.id,
-            quizRemoteBranchId: filteredRepositoryFile.parentRemoteBranchId
-          }))
-        ))
-      : []
-      const quizRemoteRepositoryFileRes = await createQuizRemoteRepositoryFile(createNewBranchRemoteRepositoryFiles.flat())
-
-      updateBranches.forEach(updateBranch => {
-        updateQuizBranch(
-          Number(updateBranch.branchId),
-          {
-            quizBranchName: updateBranch.branchName,
-            quizFirstOrLastId: Number(quizFirstOrLastId)
-          }
-        )
-      })
-      updateWorktreeFiles.forEach(updateWorktreeFile => {
-        updateQuizWorktreeFile(
-          Number(updateWorktreeFile.worktreeFileId),
-          {
-            quizWorktreeFileName: updateWorktreeFile.fileName,
-            quizWorktreeFileTextStatus: updateWorktreeFile.textStatus,
-            quizBranchId: Number(updateWorktreeFile.parentBranchId)
-          }
-        )
-      })
-      updateIndexFiles.forEach(updateIndexFile => {
-        updateQuizIndexFile(
-          Number(updateIndexFile.indexFileId),
-          {
-            quizIndexFileName: updateIndexFile.fileName,
-            quizIndexFileTextStatus: updateIndexFile.textStatus,
-            quizBranchId: Number(updateIndexFile.parentBranchId)
-          }
-        )
-      })
-
-      deleteBranches.forEach(deleteBranch => deleteQuizBranch(Number(deleteBranch.branchId)))
-      deleteCommitMessages.forEach(deleteCommitMessage => deleteQuizCommitMessage(Number(deleteCommitMessage.commitMessageId)))
-      deleteWorktreeFiles.forEach(deleteWorktreeFile => deleteQuizWorktreeFile(Number(deleteWorktreeFile.worktreeFileId)))
-      deleteIndexFiles.forEach(deleteIndexFile => deleteQuizIndexFile(Number(deleteIndexFile.indexFileId)))
-      deleteRepositoryFiles.forEach(deleteRepositoryFile => deleteQuizRepositoryFile(Number(deleteRepositoryFile.repositoryFileId)))
-      deleteFileHistoryForCansellCommits.forEach(deleteFileHistoryForCansellCommit => deleteQuizHistoryOfCommittedFile(Number(deleteFileHistoryForCansellCommit.historyFileId)))
-      deleteRemoteRepositoryFiles.forEach(deleteRemoteRepositoryFile => deleteQuizRemoteRepositoryFile(Number(deleteRemoteRepositoryFile.remoteRepositoryFileId)))
-      deleteRemoteCommitMessages.forEach(deleteRemoteCommitMessage => deleteQuizRemoteCommitMessage(Number(deleteRemoteCommitMessage.remoteCommitMessageId)))
-      deleteRemoteBranches.forEach(deleteRemoteBranch => deleteQuizRemoteBranch(Number(deleteRemoteBranch.remoteBranchId)))
-
-      //確認
-      console.log("既存ブランチからworkFile新規作成",quizWorktreeFileRes)
-      console.log("既存ブランチからindexFile新規作成",quizIndexFileRes)
-      console.log("既存ブランチからCommitMessage新規作成",QuizCommitMessageRes)
-      console.log("既存ブランチからHistoryOfCommittedFile新規作成",quizHistoryOfCommittedFileRes)
-      console.log("既存ブランチからRepositoryFile新規作成",quizRepositoryFileRes)
-      console.log("既存ブランチからRemoteCommitMessage新規作成",QuizRemoteCommitMessageRes)
-      console.log("既存ブランチからRemoteRepositoryFile新規作成",quizRemoteRepositoryFileRes)
-    } catch (err) {
-      console.log(err)
-    }
-  }
-
   useEffect(() => {
     location.pathname === (`/quiz/edit/${id}`) && handleGetQuizData()
     // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -1267,6 +776,7 @@ const CreateQuiz: React.FC = () => {
       value={{
         quizTitle, setQuizTitle,
         quizIntroduction, setQuizIntroduction,
+        quizFirstOrLastId, setQuizFirstOrLastId,
         text, setText,
         addText, setAddText,
         gitInit,setGitInit,
@@ -1302,46 +812,7 @@ const CreateQuiz: React.FC = () => {
       <AboutQuiz />
       <Terminal />
     {/* </layout> */}
-      {location.pathname === "/quiz" &&
-        <Button
-          type="submit"
-          variant="contained"
-          size="large"
-          fullWidth
-          color="default"
-          disabled={!quizTitle || !quizIntroduction || !quizType ? true : false}
-          className={classes.submitBtn}
-          onClick={
-            handleCreateQuizSubmit(
-              branches,
-              worktreeFiles,
-              indexFiles,
-              commitMessages,
-              fileHistoryForCansellCommits,
-              repositoryFiles,
-              remoteBranches,
-              remoteCommitMessages,
-              remoteRepositoryFiles
-              )
-            }
-        >
-          Submit
-        </Button>
-      }
-      {location.pathname === (`/quiz/edit/${id}`) &&
-      <Button
-          type="submit"
-          variant="contained"
-          size="large"
-          fullWidth
-          color="default"
-          disabled={!quizTitle || !quizIntroduction || !quizType ? true : false}
-          className={classes.submitBtn}
-          onClick={handleUpdateQuizData}
-        >
-          Submit
-        </Button>
-        }
+      <CreateOrUpdateQuizButton />
         <>
         {/* ↓配列の中身確認用 */}
         <p>[branch]</p>
